added a discussion of babel in the related work section
added a footnote about the attack on babel timestamps

added a short section on how using mixminion with many messages
is going to be harder than we think. it needs more work; please
go fix it.

added a few more points to the conclusion. if you have anything else
on your mind as 'future directions' or 'unsolved problems', please go
add them.